requirements.txt
torch>=1.13.0
torchvision>=0.14.0
insightface>=0.7.3
onnxruntime>=1.17.0
opencv-python>=4.7.0
numpy>=1.23.0
pillow>=9.4.0
scikit-learn>=1.2.0
scipy>=1.10.0
faiss-cpu>=1.7.3
fastapi>=0.95.0
uvicorn>=0.21.0
python-multipart>=0.0.6
psycopg2-binary>=2.9.5
redis>=4.5.0
celery>=5.2.7
pydantic>=1.10.0
gfpgan>=1.3.8
realesrgan>=0.3.0
facexlib>=0.3.0
basicsr>=1.4.2
